refactor: some name changes
